+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
-## [Unreleased]
-- Add "gem" category to all materials [#2]
+## [3.0] - 2022-05-01
+- Updated for Silent Gear 2.10+ (Minecraft 1.18.2)
+### Added
+- "gem" category to all materials [#2]
+- Tourmaline [#1]
 
 ## [2.0] - 2020-10-21
 - Updated for Silent Gear 2.3+ (may still work on Minecraft 1.15.2 as well)